fix: LikeStatus long int casting 문제 해결
Former-commit-id: 84e0300de58ae9136915a391847feb770cbbaa09

+++ backend/core/core-api/src/main/java/com/ssafy/s12p21d206/achu/api/controller/goods/GoodsDetailResponse.java
@@ -31,7 +31,7 @@ public record GoodsDetailResponse(
         goodsDetail.goods().tradeStatus(),
         goodsDetail.goods().price(),
         goodsDetail.goods().defaultDateTime().createdAt(),
-        (long) likeStatus.count(),
+        likeStatus.count(),
         likeStatus.isLike(),
         category,
         seller);
