+++ backend/core/core-domain/src/main/java/com/ssafy/s12p21d206/achu/domain/LikeStatus.java
@@ -1,3 +1,3 @@
 package com.ssafy.s12p21d206.achu.domain;
 
-public record LikeStatus(int count, boolean isLike) {}
+public record LikeStatus(long count, boolean isLike) {}
